refactor(review): extract ReviewDialogHost.Show for AF + Inspector modal lifecycle (F24)
